backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
qdrant-client==1.12.0
asyncpg==0.30.0
pydantic>=2.12.3
pydantic-settings==2.6.0
openai-agents==0.6.3
google-generativeai==0.8.4
python-dotenv==1.0.1
psycopg[binary,pool]==3.2.3
sqlalchemy==2.0.35
alembic==1.13.3
uuid7==0.1.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pytest==8.3.3
pytest-asyncio==0.23.7
httpx==0.27.2
python-slugify==8.0.4